New API End point for BusinessArea

+++ AtoCash/Controllers/BasicControlrs/ExpenseCategoriesController.cs
@@ -14,7 +14,7 @@ namespace AtoCash.Controllers
 {
     [Route("api/[controller]/[Action]")]
     [ApiController]
-   [Authorize(Roles = "AtominosAdmin, Admin, Manager, Finmgr, User")]
+  // [Authorize(Roles = "AtominosAdmin, Admin, Manager, Finmgr, User")]
     public class ExpenseCategoriesController : ControllerBase
     {
         private readonly AtoCashDbContext _context;
@@ -96,7 +96,7 @@ namespace AtoCash.Controllers
 
         // PUT: api/ExpenseCategories/5
         [HttpPut("{id}")]
-       [Authorize(Roles = "AtominosAdmin, Admin, Manager, Finmgr")]
+      // [Authorize(Roles = "AtominosAdmin, Admin, Manager, Finmgr")]
         public async Task<IActionResult> PutExpenseCategory(int id, ExpenseCategoryDTO expenseCategoryDTO)
         {
             if (id != expenseCategoryDTO.Id)
@@ -128,7 +128,7 @@ namespace AtoCash.Controllers
         // POST: api/ExpenseCategories
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPost]
-       [Authorize(Roles = "AtominosAdmin, Admin, Manager, Finmgr")]
+      // [Authorize(Roles = "AtominosAdmin, Admin, Manager, Finmgr")]
         public async Task<ActionResult<ExpenseCategory>> PostExpenseCategory(ExpenseCategoryDTO expenseCategoryDTO)
         {
             var eCategory = _context.ExpenseCategories.Where(e => e.ExpenseCategoryName == expenseCategoryDTO.ExpenseCategoryName).FirstOrDefault();
@@ -149,7 +149,7 @@ namespace AtoCash.Controllers
 
         // DELETE: api/ExpenseCategories/5
         [HttpDelete("{id}")]
-       [Authorize(Roles = "AtominosAdmin, Admin, Manager, Finmgr")]
+      // [Authorize(Roles = "AtominosAdmin, Admin, Manager, Finmgr")]
         public async Task<IActionResult> DeleteExpenseCategory(int id)
         {
 
@@ -159,13 +159,13 @@ namespace AtoCash.Controllers
                 return Conflict(new RespStatus { Status = "Failure", Message = "Expense-Category Id Invalid!" });
             }
 
-            /* Commented as of now, till we create a Column in Expense Sub Claim Table */ 
-            /*var expReimburse = _context.ExpenseSubClaims.Where(d => d.ExpenseCategoryId == id).FirstOrDefault();
+            //Commented as of now, till we create a Column in Expense Sub Claim Table */ 
+            var expType = _context.ExpenseTypes.Where(d => d.ExpenseCategoryId == id).FirstOrDefault();
 
-            if (expReimburse != null)
+            if (expType != null)
             {
-            */    return Conflict(new RespStatus { Status = "Failure", Message = "Expense-Category in use for Expense Reimburse!" });
-            /*}*/
+                return Conflict(new RespStatus { Status = "Failure", Message = "Expense-Category in use for Expense Type!" });
+            }
 
             _context.ExpenseCategories.Remove(expenseCategory);
             await _context.SaveChangesAsync();
